Moved classes and hooks to a separate file
